Overhauled prompt expansions and shell options
Prompt expansion now uses a simple tokenizer/parser with recursive descent

Recursive descent is added because of the addition of 'hide groups'.

hide groups are escape sequences which contain other escape sequences, like this: \(\u\)

If none of the escape sequences inside of the hide group expand into anything, then none of the text in the hide group is displayed.

Shell options are now contained in the shopt.rs module, and are accessed using a json-like syntax, i.e. 'getopt prompt.git_signs.branch_icon'